adicionado método .intValue na variável do tipo Integer em registroPagamento model e aluno model

+++ TF-Academia/Backend/academia/src/main/java/com/timedev/academia/model/Aluno.java
@@ -80,7 +80,7 @@ public class Aluno extends Pessoa{
 	public void removerRegistro(Integer id) {
 		RegistroPagamento recebeRegistro = null;
 		for (RegistroPagamento registroPagamento : registrosDePagamentos) {
-			if(registroPagamento.getId() == id) {
+			if(registroPagamento.getId() == id.intValue()) {
 				recebeRegistro = registroPagamento; 
 			}
 		}
